requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
csscompressor==0.9.5
django-appconf==1.0.3     # via django-compressor, django-imagefit
django-compressor==2.3
django-extensions==2.2.5
django-imagefit==0.6.2
django-libsass==0.7
django-markdownx==2.0.28
django-tagulous==0.14.1
django==2.2.6
gunicorn==19.9.0
libsass==0.19.3           # via django-libsass
markdown==3.1.1           # via django-markdownx
pillow==6.2.1
psycopg2==2.8.4
pytz==2019.3              # via django
rcssmin==1.0.6            # via django-compressor
rjsmin==1.1.0             # via django-compressor
six==1.12.0               # via django-appconf, django-extensions, libsass
sqlparse==0.3.0           # via django

# The following packages are considered to be unsafe in a requirements file:
# pip==19.3.1               # via django-markdownx
# setuptools==41.4.0        # via markdown
